pyproject.toml
[project]
name = "python-harness"
version = "0.0.1"
description = "An agentic harness tool for Python codebase understanding and hard evaluation."
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "ty>=0.0.1", # Assuming ty is available or will be replaced with actual LSP integration
    "radon>=6.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["python_harness", "python_harness.*"]

[project.scripts]
harness = "python_harness.cli:app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    "vendors/.*"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=python_harness --cov-report=term-missing --cov-report=html"
testpaths = [
    "tests",
]

